feat: Edited examples per Michael's design and docs added
