+++ modules/react/layout/stories/Grid.stories.mdx
@@ -2,9 +2,9 @@ import {Meta} from '@storybook/addon-docs';
 
 import {Grid} from '@workday/canvas-kit-react/layout';
 // examples
-import {GridCard} from './examples/Grid/GridCard';
 import {GridLayout} from './examples/Grid/GridLayout';
 import {UIExample} from './examples/Grid/UIExample';
+import {GridLayoutInteractive} from './examples/Grid/GridLayoutInteractive';
 import {GridStyle} from './examples/PropTables.splitprops.tsx';
 
 <Meta title="Components/Containers/Layout/Grid" component={Grid} />
@@ -30,27 +30,28 @@ yarn add @workday/canvas-kit-react
 You can nest `Grid` components to build layouts with rows and columns. You can also nest `Grid`
 components within `<Grid.Item/>` components for even more possibilities.
 
-In the below example, we have a `Grid` container which has a green border. It contains nine elements
-with one of them being grey. This is a `<Grid.Item/>` component and has access to child specific
-props, the `<Grid/>` component only has access to container props. This is to show what the
-`<Grid.Item/>` component can do. **This is NOT best practice for accessibility reasons and we
-suggest against following this. This is merely to showcase the child specific props for the
-`<Grid.Item/>` component.**
+In the below example, we have a `Grid` container. It contains nine elements with one of them being
+grey. This is a `<Grid.Item/>` component and has access to child specific props, the `<Grid/>`
+component only has access to container props. This is to show what the `<Grid.Item/>` component can
+do.
 
 In the layout, the `<Grid.Item/>` component (grey element), has a `Row` and `Col` button. These
-buttons are interactive and if you click the `Row` button, it will move the element between the rows
-and if you click the `Col` button, it will move the element through the columns. The `<Grid.Item/>`
-component will move among the parent grid when the `Row` or `Col` button is clicked. These buttons
-are updating `gridRowStart`(`grid-row-start` css grid property) and
+buttons are interactive and if you click the `Row` button, it will move the `<Grid.Item/>` element
+between the rows and if you click the `Col` button, it will move it through the columns. These
+buttons are updating `gridRowStart`(`grid-row-start` css grid property) and
 `gridColumnStart`(`grid-column-start` css grid property). This is to showcase the child specific
 props for the `<Grid.Item/>` component.
 
-In the second example below, we have a `Grid` container which has a red-orange border. It contains
-two rows with one having seven elements and the other has two elements. This example is to show how
-a user can use the `<Grid.Item/>` component to wrap a `<Grid/>` component. This would allow a user
-to have access to the child props of `<Grid.Item/>` to place a layout where needed. The row with
-elements 3 - 7 is put before the 1 & 2 elements since that layout is wrapped by the `<Grid.Item/>`
-component.
+<ExampleCodeBlock code={GridLayoutInteractive} />
+
+**This is NOT best practice for accessibility reasons and we suggest against following this. This is
+merely to showcase the child specific props for the `<Grid.Item/>` component.**
+
+In the second example below, we have a `Grid` container. It contains two rows with one having seven
+elements and the other has two elements. This example is to show how a user can use the
+`<Grid.Item/>` component to wrap a `<Grid/>` component. This would allow a user to have access to
+the child props of `<Grid.Item/>` to place a layout where needed. The row with elements 3 - 7 is put
+before the 1 & 2 elements since that layout is wrapped by the `<Grid.Item/>` component.
 
 <ExampleCodeBlock code={GridLayout} />
 
@@ -63,13 +64,6 @@ how `gridArea` works with `gridTemplateAreas`,
 [see here](https://developer.mozilla.org/en-US/docs/Web/CSS/grid-template-areas).
 
 <ExampleCodeBlock code={UIExample} />
-
-<!-- ### Small Containers
-
-You can also use `Grid` to create layouts within smaller containers, such as this card example
-below:
-
-<ExampleCodeBlock code={GridCard} /> -->
 
 ### Grid vs. Box
