--- modules/react/layout/stories/examples/Grid/Example.tsx
@@ -1,81 +1,0 @@
-import * as React from 'react';
-import {Grid, GridProps, Box} from '@workday/canvas-kit-react/layout';
-import {type} from '@workday/canvas-kit-react/tokens';
-
-// temporary placeholder until type components are added to canvas-kit
-const H3 = props => (
-  <h3
-    style={{
-      ...type.levels.body.large,
-      ...type.variants.inverse,
-      margin: 0,
-      fontWeight: type.properties.fontWeights.bold,
-    }}
-    {...props}
-  />
-);
-
-const Body = props => (
-  <p style={{...type.levels.body.small, ...type.variants.inverse, margin: 0}} {...props} />
-);
-
-const Card = ({children, ...props}: GridProps) => (
-  <Grid
-    backgroundColor="frenchVanilla100"
-    borderRadius="m"
-    depth={1}
-    gridAutoFlow="row"
-    margin="xs"
-    padding="s"
-    {...props}
-  >
-    {children}
-  </Grid>
-);
-
-export const Example = () => {
-  return (
-    <Grid as="section" padding="s">
-      <h2>Canvas Principles</h2>
-      <Grid gridTemplateColumns="repeat(auto-fit, minmax(320px, 1fr))">
-        <Card backgroundColor="blueberry400">
-          <H3>Empower over Enforce</H3>
-          <Box paddingY="xs">
-            <Body>
-              Encourage our user's expression. Stay out of the way and provide them with the tools
-              and resources to build their vision.
-            </Body>
-          </Box>
-        </Card>
-        <Card backgroundColor="juicyPear500">
-          <H3>Evolution over Perfection</H3>
-          <Box paddingY="xs">
-            <Body>
-              Nothing is ever perfect – embrace that. Make educated assumptions, validate and test
-              our decisions, then iterate! Aim of continous rather than perfect solutions.
-            </Body>
-          </Box>
-        </Card>
-        <Card backgroundColor="chiliMango400">
-          <H3>Simple over Clever</H3>
-          <Box paddingY="xs">
-            <Body>
-              Simple solutions invite the user in - clever solutions invite complexity. Make the
-              system easy and predictable, and progressively disclose advanced functionality.
-            </Body>
-          </Box>
-        </Card>
-        <Card backgroundColor="blackberry400">
-          <H3>Everyone over Every One</H3>
-          <Box paddingY="xs">
-            <Body>
-              Each piece of the system is designs and built to be accessible, while still providing
-              the best experience for all consumers. But not if something is focused on a single use
-              case and negates the usability for others.
-            </Body>
-          </Box>
-        </Card>
-      </Grid>
-    </Grid>
-  );
-};
